academic-advisor-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
firebase-admin==6.2.0
PyPDF2==3.0.1
PyMuPDF==1.24.7
python-docx==1.1.0
docx2txt==0.8
pytesseract==0.3.10
Pillow==10.1.0
spacy==3.7.5
nltk==3.8.1
transformers==4.36.2
torch==2.3.1+cpu
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.3
textblob==0.18.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.1
redis==5.0.1
celery==5.3.6
sentry-sdk==1.40.0
prometheus-fastapi-instrumentator==6.1.0
boto3==1.34.131
psycopg2-binary==2.9.9
httpx==0.27.0
pytest==7.4.4
pytest-asyncio==0.23.6
black==24.4.2
flake8==7.0.2